Update style.css
The error was simple, I have now added the 
float:left;

to the Div class of each image and text. Upon adding the " padding-top:500px; " I realised that the image would move but the text would remain leading me to believe that the text was causing the error. Since I only had the float on the image id's I then added it to the div class so the text would float alongside the image.